+++ src/main/java/com/ayakacraft/authlibproxyforserver/utils/mixin/AuthProxyMixinPlugin.java
@@ -1,0 +1,51 @@
+/*
+ * This file is part of the AuthlibProxyForServer project, licensed under the
+ * GNU General Public License v3.0
+ *
+ * Copyright (C) 2025  Calboot and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.ayakacraft.authlibproxyforserver.utils.mixin;
+
+import com.ayakacraft.authlibproxyforserver.AuthlibProxyForServer;
+import me.fallenbreath.conditionalmixin.api.mixin.RestrictiveMixinConfigPlugin;
+
+import java.util.List;
+import java.util.Set;
+
+public class AuthProxyMixinPlugin extends RestrictiveMixinConfigPlugin {
+
+    @Override
+    protected void onRestrictionCheckFailed(String mixinClassName, String reason) {
+        AuthlibProxyForServer.LOGGER.debug("Disabled mixin {} due to {}", mixinClassName, reason);
+    }
+
+    @Override
+    public String getRefMapperConfig() {
+        return null;
+    }
+
+    @Override
+    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
+
+    }
+
+    @Override
+    public List<String> getMixins() {
+        return List.of();
+    }
+
+}