+++ versions/1.15.2/src/main/java/com/ayakacraft/authlibproxyforserver/mixin/YggdrasilGameProfileRepositoryAccessor.java
@@ -1,0 +1,38 @@
+/*
+ * This file is part of the AuthlibProxyForServer project, licensed under the
+ * GNU General Public License v3.0
+ *
+ * Copyright (C) 2025  Calboot and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.ayakacraft.authlibproxyforserver.mixin;
+
+import com.mojang.authlib.yggdrasil.YggdrasilGameProfileRepository;
+import me.fallenbreath.conditionalmixin.api.annotation.Condition;
+import me.fallenbreath.conditionalmixin.api.annotation.Restriction;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Restriction(require = @Condition(value = "minecraft", versionPredicates = "<1.16"))
+@Mixin(value = YggdrasilGameProfileRepository.class, remap = false)
+public interface YggdrasilGameProfileRepositoryAccessor {
+
+    @Accessor("BASE_URL")
+    static String getBaseUrl() {
+        throw new AssertionError();
+    }
+
+}