+++ changelog.md
@@ -1,3 +1,4 @@
 ### General
 
+- ***Added*** command `/authproxy ping`
 - Support for Minecraft 1.21.7